Updated some stuff
