+++ coverage.sh
@@ -27,11 +27,11 @@ function quit() {
 }
 # change to example directory for execution (because it uses hardcoded filepaths, and the testable
 # examples don't work when executed outside of that directory
-go test -c -coverpkg=$pck1 -covermode=atomic -o "$root_res" $pck1
+go test --tags=evolution_test -c -coverpkg=$pck1 -covermode=atomic -o "$root_res" $pck1
 # convert binary to go formatted
-go tool test2json -t "$root_res" -test.v -test.coverprofile "$root_out"
+go tool  test2json -t "$root_res" -test.v -test.coverprofile "$root_out"
 
-go test -c -coverpkg=$pck1 -covermode=atomic -o "$source_res" $pck2
+go test --tags=evolution_test -c -coverpkg=$pck1 -covermode=atomic -o "$source_res" $pck2
 go tool test2json -t "$source_res" -test.v -test.coverprofile "$source_out"
 
 # delete aggregate file
